backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pandas==2.2.2
duckdb==1.0.0
pydantic==2.7.4
pydantic-settings==2.3.4
python-multipart==0.0.9
usaddress==0.5.10
uszipcode==1.0.1
openpyxl==3.1.5
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.0
pytest==8.2.2
pytest-asyncio==0.23.7
faker==25.8.0
factory-boy==3.3.0